refactor: 컴포넌트를 배럴 파일에도 추가
