+++ src/components/index.js
@@ -1,3 +1,6 @@
 export * from "./header";
 export * from "./footer";
 export * from "./nav";
+export * from "./posts";
+export * from "./profileForm";
+export * from "./loginForm";
